moved code to open api struct
